Implement label replacement.
We want to restrict ourselves to three labels in the training, so
paragraphs that do not already have one of the labels we want as top
label need to get our default label.

Equality for labels needs to be able to handle None, because there may
be no top label.