Refine end-of-paragraph.
Some of the weirder behaviors so far have come from trying to make
sure that taxonomic paragraphs end correctly without pulling in
unrelated components. It is for this reason we end a paragraph on a
period at end-of-line.

We've gotten more sophisticated about deciding if the dot at the end
of the line is really a period or just an indication of an
abbreviation. But we can't skip just any abbreviation--e.g. "nov." is
a common abbreviation at the end of a taxonomic paragraph. I
hand-built a list of abbreviations from Mycotaxon Vol056 which do not
end a Taxonomy. This really wants an ML solution.

New document now triggers a new paragraph.

Synonymy like "(Syn. foo bar)" ends a paragraph.

Year like "(1945)" ends a paragraph.

Add a bunch of arguments that control output.