New default classifier, paragraph ends with :, dispose of large objects.
I've been reading the best classifier incorrectly. The best for base3
(the current base) is CalibratedClassifierCV.

A taxon occasionally starts after a : at the end of a line.

I'm running out of memory now trying to calculate
KNeighborsClassifier. Deleting the earlier phases did not help, but is
probably a good idea anyway.