Add line number handling.
We introduce a File object which handles each file separately. This
allows us to eliminate the class-global line count in Line.

We have both page numbers by counting form feeds and page numbers read
from the first line of a page. Handling of page numbers at the bottom
is a task for the future.

The former is valuable for linking directly to a page in a pdf. The
latter is useful for finding an article in a bibliographic reference.