Explore brute-force recognition of nomenclature.
nomenclatures.txt is an assortment of nomenclatures, mostly the full
notation I want to find, but a few that are short of critical
components.

nomenclature.py is a program that tries to recognize nomenclatures
using supposedly complete lists of species (MycoBank), author names &
abbreviations (MycoBank and Index Fungorum), and ISO 4 journal names.

My conclusions:

Brute force works better than the machine learning methods I've used
so far, but The code is crappy and still has a lot defects--it needs a
ground-up redesign and reimplementation.

I was able to build a regular expression to find the end of a
nomenclature of the desired type. I still don't have a great way to
find the start of a multi-line nomenclature.